Sleep

+++ SleepMode.X/main.c
@@ -35,9 +35,7 @@ void __interrupt() INT_TMR0(void){
         PIR1bits.ADIF = 0;
         ADCON0bits.ADON = 0;
         
-        int *B = BinTOBcd(((ADRESH<<8) + ADRESL));
-        //int *B = BinTOBcd(iADC);
-        //int B[3] = {7, 8, 10};
+        unsigned short *B = BinTOBcd((unsigned long)((ADRESH<<8) + ADRESL));
         UART_print(ASCII_Con(B[2], B[1], B[0]));
         return;
     }
